+++ packages/storybook-html/stories/components/alert/alert.stories.js
@@ -71,6 +71,90 @@ export const alertWithTitle = {
   },
 };
 
+export const alertWithHeading = {
+  render: () => {
+    return `
+    <div class="uq-alert uq-alert--info" role="alert">
+      <div class="uq-alert__container">
+        <div class="uq-alert__message">
+          <h1>Heading 1</h1>
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+          <a href="#" title="Link" class="uq-button--inline">Link</a>
+        </div>
+        <button class="uq-alert__close">close</button>
+      </div>
+    </div>
+
+    <br />
+    
+    <div class="uq-alert uq-alert--info" role="alert">
+      <div class="uq-alert__container">
+        <div class="uq-alert__message">
+          <h2>Heading 2</h2>
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+          <a href="#" title="Link" class="uq-button--inline">Link</a>
+        </div>
+        <button class="uq-alert__close">close</button>
+      </div>
+    </div>
+
+    <br />
+
+    <div class="uq-alert uq-alert--info" role="alert">
+      <div class="uq-alert__container">
+        <div class="uq-alert__message">
+          <h3>Heading 3</h3>
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+          <a href="#" title="Link" class="uq-button--inline">Link</a>
+        </div>
+        <button class="uq-alert__close">close</button>
+      </div>
+    </div>
+
+    <br />
+
+    <div class="uq-alert uq-alert--info" role="alert">
+      <div class="uq-alert__container">
+        <div class="uq-alert__message">
+          <h4>Heading 4</h4>
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+          <a href="#" title="Link" class="uq-button--inline">Link</a>
+        </div>
+        <button class="uq-alert__close">close</button>
+      </div>
+    </div>
+
+    <br />
+
+    <div class="uq-alert uq-alert--info" role="alert">
+      <div class="uq-alert__container">
+        <div class="uq-alert__message">
+          <h5>Heading 5</h5>
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+          <a href="#" title="Link" class="uq-button--inline">Link</a>
+        </div>
+        <button class="uq-alert__close">close</button>
+      </div>
+    </div>
+
+    <br />
+
+    <div class="uq-alert uq-alert--info" role="alert">
+      <div class="uq-alert__container">
+        <div class="uq-alert__message">
+          <h6>Heading 6</h6>
+          <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.</p>
+          <a href="#" title="Link" class="uq-button--inline">Link</a>
+        </div>
+        <button class="uq-alert__close">close</button>
+      </div>
+    </div>
+    `;
+  },
+
+  name: "Headings",
+};
+
 export const infoGlobal = {
   render: () => {
     return `